feat(athena): better error output
Ref: #2115
